Split profile DTOs into request and response subfolders and add ProfileDto response DTO
