+++ src/modules/profile/profile.controller.ts
@@ -1,11 +1,13 @@
 import { Body, Controller, Get, Param, Patch, UseGuards } from '@nestjs/common';
-import { CreateProfileDto } from './dto/create-profile.dto';
 import { ProfileService } from './profile.service';
 import { AuthGuard } from 'src/core/guards/auth.guard';
 import { CurrentUser } from 'src/common/decorators/current-user.decorator';
 import { User } from 'src/database/entities';
-import { UpdateProfileDto } from './dto/update-profile.dto';
+import { UpdateProfileDto } from './dto/request/update-profile.dto';
+import { Serialize } from 'src/core/interceptors/serialize.interceptor';
+import { ProfileDto } from './dto/response/profile.dto';
 
+@Serialize(ProfileDto)
 @Controller('profile')
 export class ProfileController {
     constructor(private profileService: ProfileService) { }
